Refactored login and register to authService.js
